# backend/spec/openapi.yaml
openapi: "3.1.1"
info:
  title: My Library API
  description: |
    Book library management system.
    All endpoints return JSON. Some endpoints require authentication
    (Bearer token from `/api/auth/login`).
  version: "0.3.0"
  contact:
    name: My Library
servers:
  - url: http://127.0.0.1:8000
    description: Local development
  - url: https://dingfengbo.top/my-library
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookCreation:
      type: object
      required: [isbn, title]
      properties:
        isbn: { type: string }
        title: { type: string }
        title_cn: { type: string }
        author_ids: { type: array, items: { type: integer } }
        translator: { type: string, nullable: true }
        publisher_id: { type: integer, nullable: true }
        publish_date: { type: string, nullable: true }
        brand_id: { type: integer, nullable: true }
        book_series_id: { type: integer, nullable: true }
        binding_type: { type: string, nullable: true }
        paper_type: { type: string, nullable: true }
        pages: { type: integer, nullable: true }
        book_count: { type: integer, nullable: true }
        language: { type: string, nullable: true }
        compose_type: { type: string, nullable: true }
        price: { type: number, nullable: true }
        purchase_price: { type: number, nullable: true }
        purchase_date: { type: string, nullable: true }
        thumb_image: { type: string, nullable: true }
        link: { type: string, nullable: true }
        category_id: { type: integer, nullable: true }
        bookshelf_id: { type: integer, nullable: true }
        read_state: { type: string, nullable: true }
        catalog: { type: string, nullable: true }
        introduction: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        registered: { type: boolean, nullable: true }
        edition: { type: string, nullable: true }
        printing_info: { type: string, nullable: true }
        printed_number: { type: integer, nullable: true }
        douban_score: { type: number, nullable: true }
        purchase_store: { type: string, nullable: true }
        tags: { type: array, nullable: true, items: { type: string } }
        in_wish: { type: boolean }
    BookUpdate:
      type: object
      properties:
        isbn: { type: string }
        title: { type: string }
        title_cn: { type: string }
        author_ids: { type: array, items: { type: integer } }
        translator: { type: string, nullable: true }
        publisher_id: { type: integer, nullable: true }
        publish_date: { type: string, nullable: true }
        brand_id: { type: integer, nullable: true }
        book_series_id: { type: integer, nullable: true }
        binding_type: { type: string, nullable: true }
        paper_type: { type: string, nullable: true }
        pages: { type: integer, nullable: true }
        book_count: { type: integer, nullable: true }
        language: { type: string, nullable: true }
        compose_type: { type: string, nullable: true }
        price: { type: number, nullable: true }
        purchase_price: { type: number, nullable: true }
        purchase_date: { type: string, nullable: true }
        thumb_image: { type: string, nullable: true }
        link: { type: string, nullable: true }
        category_id: { type: integer, nullable: true }
        bookshelf_id: { type: integer, nullable: true }
        read_state: { type: string, nullable: true }
        catalog: { type: string, nullable: true }
        introduction: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        registered: { type: boolean, nullable: true }
        edition: { type: string, nullable: true }
        printing_info: { type: string, nullable: true }
        printed_number: { type: integer, nullable: true }
        douban_score: { type: number, nullable: true }
        purchase_store: { type: string, nullable: true }
        tags: { type: array, nullable: true, items: { type: string } }
        in_wish: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BookResponse:
      type: object
      properties:
        id: { type: integer }
        isbn: { type: string }
        title: { type: string }
        title_cn: { type: string }
        authors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
        translator: { type: string, nullable: true }
        publisher: { $ref: "#/components/schemas/PublisherSimple", nullable: true }
        publish_date: { type: string, format: date-time, nullable: true }
        brand: { $ref: "#/components/schemas/BrandSimple", nullable: true }
        book_series: { $ref: "#/components/schemas/SeriesSimple", nullable: true }
        binding_type: { type: string, nullable: true }
        paper_type: { type: string, nullable: true }
        pages: { type: integer, nullable: true }
        book_count: { type: integer, nullable: true }
        language: { type: string, nullable: true }
        compose_type: { type: string, nullable: true }
        price: { type: number, nullable: true }
        purchase_price: { type: number, nullable: true }
        purchase_date: { type: string, format: date-time, nullable: true }
        thumb_image: { type: string, nullable: true }
        link: { type: string, nullable: true }
        category: { $ref: "#/components/schemas/CategorySimple", nullable: true }
        bookshelf: { $ref: "#/components/schemas/BookshelfSimple", nullable: true }
        read_state: { type: string, nullable: true }
        catalog: { type: string, nullable: true }
        introduction: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        registered: { type: boolean, nullable: true }
        edition: { type: string, nullable: true }
        printing_info: { type: string, nullable: true }
        printed_number: { type: integer, nullable: true }
        douban_score: { type: number, nullable: true }
        purchase_store: { type: string, nullable: true }
        tags: { type: array, nullable: true, items: { type: string } }
        in_wish: { type: boolean, nullable: true }

    # ── Simple ref types ──
    PublisherSimple:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
    BrandSimple:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
    SeriesSimple:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
    CategorySimple:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
    BookshelfSimple:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }

    # ── Author ──
...
    BookSimple:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        title_cn: { type: string, nullable: true }
        thumb_image: { type: string, nullable: true }
        isbn: { type: string, nullable: true }
        authors:
          type: array
          nullable: true
          items: { type: string }

    # ── Publisher ──
...
    AddBookToCollection:
      type: object
      required: [book_id]
      properties:
        book_id: { type: integer }
...
    BookCollectionResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        intro: { type: string, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        total_books: { type: integer, nullable: true }
        books:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/BookSimple"
...
    AddBookToPlan:
      type: object
      required: [book_id]
      properties:
        book_id: { type: integer }
    ReadingPlanBookSimple:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        title_cn: { type: string, nullable: true }
        thumb_image: { type: string, nullable: true }
        isbn: { type: string, nullable: true }
        authors:
          type: array
          nullable: true
          items: { type: string }
        read_state: { type: string, nullable: true }
        book_count: { type: integer, nullable: true }
...
    ReadingPlanResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        intro: { type: string, nullable: true }
        start_date: { type: string, format: date, nullable: true }
        end_date: { type: string, format: date, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        total_books: { type: integer, nullable: true }
        progress: { type: number, nullable: true }
        books:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ReadingPlanBookSimple"
...
    HTTPError:
      type: object
      properties:
        detail: { type: string }
paths:
...
  /api/books:
    get:
      tags: [Books]
      summary: List books (paginated)
      description: Returns books with in_wish=false only.
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: sort_by, in: query, schema: { type: string, default: title, enum: [title, created_at] } }
        - { name: isbn, in: query, schema: { type: string } }
        - { name: title, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
        - { name: publisher, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: min_price, in: query, schema: { type: number } }
        - { name: max_price, in: query, schema: { type: number } }
      responses:
        "200":
          description: Paginated book list
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookResponse"
                  total_pages: { type: integer }
                  total_books: { type: integer }
    post:
      tags: [Books]
      summary: Create a new book
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookCreation"
      responses:
        "200":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
  /api/books/search:
    get:
      tags: [Books]
      summary: Search books with filters
      description: Returns books with in_wish=false only.
      parameters:
        - { name: isbn, in: query, schema: { type: string } }
        - { name: title, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
        - { name: publisher, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: min_price, in: query, schema: { type: number } }
        - { name: max_price, in: query, schema: { type: number } }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
...
  /api/books/{book_id}:
    get:
      tags: [Books]
      summary: Get book details
      parameters:
        - { name: book_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "404":
          description: Book not found
    put:
      tags: [Books]
      summary: Update a book
      security:
        - BearerAuth: []
      parameters:
        - { name: book_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
    delete:
      tags: [Books]
      summary: Delete a book
      security:
        - BearerAuth: []
      parameters:
        - { name: book_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /api/books/{book_id}/similar:
    get:
      tags: [Books]
      summary: Get similar books
      parameters:
        - { name: book_id, in: path, required: true, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, default: 5 } }
      responses:
        "200":
          description: Similar books
          content:
            application/json:
              schema:
                type: object
                properties:
                  similar_books:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookSimple"
                  total_similar_books: { type: integer }

  # ═══════════════════════════════════════════════
  # Authors
  # ═══════════════════════════════════════════════
...
  /api/book-collections/{collection_id}/books:
    post:
      tags: [Book Collections]
      summary: Add a single book to collection
      security:
        - BearerAuth: []
      parameters:
        - { name: collection_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBookToCollection"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookCollectionResponse"
        "400":
          description: Book already in collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "404":
          description: Book or collection not found
...
  /api/book-collections/{collection_id}/books/{book_id}:
    delete:
      tags: [Book Collections]
      summary: Remove a book from collection
      security:
        - BearerAuth: []
      parameters:
        - { name: collection_id, in: path, required: true, schema: { type: integer } }
        - { name: book_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookCollectionResponse"
        "400":
          description: Book not in collection


  # ═══════════════════════════════════════════════
  # Reading Plans
  # ═══════════════════════════════════════════════
...
  /api/reading-plans/{plan_id}/books:
    post:
      tags: [Reading Plans]
      summary: Add a single book to plan
      security:
        - BearerAuth: []
      parameters:
        - { name: plan_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBookToPlan"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingPlanResponse"
        "400":
          description: Book already in plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "404":
          description: Book or plan not found
...
  /api/reading-plans/{plan_id}/books/{book_id}:
    delete:
      tags: [Reading Plans]
      summary: Remove a book from plan
      security:
        - BearerAuth: []
      parameters:
        - { name: plan_id, in: path, required: true, schema: { type: integer } }
        - { name: book_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingPlanResponse"
        "400":
          description: Book not in plan
  # ═══════════════════════════════════════════════
  # Applications
  # ═══════════════════════════════════════════════
...